Drop FindPython3.cmake only for 1.3.x

+++ recipes/open62541/all/conanfile.py
@@ -346,7 +346,8 @@ class Open62541Conan(ConanFile):
     def _patch_sources(self):
         for patch in self.conan_data.get("patches", {}).get(self.version, []):
             tools.patch(**patch)
-        os.unlink(os.path.join(self._source_subfolder, "tools", "cmake", "FindPython3.cmake"))
+        if tools.Version(self.version) >= "1.3.1":
+            os.unlink(os.path.join(self._source_subfolder, "tools", "cmake", "FindPython3.cmake"))
 
     def build(self):
         self._patch_sources()
